fix arm64 compile

+++ havana/havok/hk_physics/constraint/ragdoll/ragdoll_constraint.cpp
@@ -32,7 +32,7 @@ class hk_Ragdoll_Constraint_Work
 			return addr;
 		}
 #else
-		static inline void *operator new (unsigned int size, void *addr){
+		static inline void *operator new (size_t size, void *addr){
 			return addr;
 		}
