+++ ivp_intern/ivp_controller_phantom.cxx
@@ -54,7 +54,7 @@ void IVP_Controller_Phantom::mindist_entered_volume(class IVP_Mindist *mindist)
 		{
 			other_object = mindist->get_synapse(0)->get_object();
 		}
-		int x = (int)mindist_object_counter->find_elem( other_object );
+        intp x = (intp)mindist_object_counter->find_elem( other_object );
 		if (x)
 		{
 			mindist_object_counter->change_elem( other_object, (void *)(x+1));
@@ -77,7 +77,7 @@ void IVP_Controller_Phantom::mindist_entered_volume(class IVP_Mindist *mindist)
 			other_object = mindist->get_synapse(0)->get_object();
 		}
 		IVP_Core *other_core= other_object->get_core();
-		int x = (int)mindist_core_counter->find_elem( other_core );
+        intp x = (intp)mindist_core_counter->find_elem( other_core );
 		if (x)
 		{
 			mindist_core_counter->change_elem( other_core, (void *)(x+1));
@@ -115,7 +115,7 @@ void IVP_Controller_Phantom:: mindist_left_volume(class IVP_Mindist *mindist)
 		{
 			other_object = mindist->get_synapse(0)->get_object();
 		}
-		int x = (int)mindist_object_counter->find_elem( other_object );
+        intp x = (intp)mindist_object_counter->find_elem( other_object );
 		if (x>1)
 		{
 			mindist_object_counter->change_elem( other_object, (void *)(x-1));
@@ -138,7 +138,7 @@ void IVP_Controller_Phantom:: mindist_left_volume(class IVP_Mindist *mindist)
 			other_object = mindist->get_synapse(0)->get_object();
 		}
 		IVP_Core *other_core= other_object->get_core();
-		int x = (int)mindist_core_counter->find_elem( other_core );
+        intp x = (intp)mindist_core_counter->find_elem( other_core );
 		if (x>1)
 		{
 			mindist_core_counter->change_elem( other_core, (void *)(x-1));
@@ -191,7 +191,7 @@ set_of_mindists(16)
 	
 	if (templat->dont_check_for_unmoveables)
 	{
-		// don´t change anything here
+		// donï¿½t change anything here
 		
 	}
 	else
