+++ ivp_surface_manager/ivp_compact_grid.hxx
@@ -67,7 +67,7 @@ public:
      *  Description:    INTERNAL METHOD
      *****************************************************************************/
     const IVP_Compact_Grid_Element *get_grid_elements() {
-    	char *base = (char *)(((int)this) + this->offset_grid_elements);
+    	char *base = (char *)(((intp)this) + this->offset_grid_elements);
 	return((const IVP_Compact_Grid_Element *)base);
     }
 
@@ -76,7 +76,7 @@ public:
      *  Description:    INTERNAL METHOD
      *****************************************************************************/
     const IVP_Compact_Ledge *get_compact_ledge_at(int i) {
-    	char *base = (char *)(((int)this) + this->offset_compact_ledge_array[i]);
+    	char *base = (char *)(((intp)this) + this->offset_compact_ledge_array[i]);
 	return((const IVP_Compact_Ledge *)base);
     }
