+++ ivp_utility/ivu_memory.hxx
@@ -105,7 +105,7 @@ inline void    *IVP_U_Memory::get_mem(unsigned int groesse)
 	    return ((void *) this->neuer_sp_block(groesse));
 	} else {
 	    speicherbeginn = p;
-		IVP_IF( ((int)op > 0x780000 ) && ((int)op < 0x792f48)) {
+		IVP_IF( ((intp)op > 0x780000 ) && ((intp)op < 0x792f48)) {
 			op++;
 			op--;
 		}
